test(hierarchical-grid): move test to virtualization spec
